# Optimize (de)compression and fix wildcopy overread