… lib/compress/zstd_opt.c
… lib/decompress/zstd_decompress_block.c
… tests/decodecorpus.c