* Bump `WILDCOPY_OVERLENGTH` to 32 to fix the wildcopy overread introduced in #1756.
* Optimize `ZSTD_wildcopy()` by removing unnecessary branches, unrolling the loop, and bumping to a 32 byte copy in stripes of 16.
* Extract `ZSTD_overlapCopy8()` into its own function.
* Add `ZSTD_safecopy()` for `ZSTD_execSequenceEnd()`. It is
  optimized for single long sequences, since that is the important
  case that can end up in `ZSTD_execSequenceEnd()`. Without this
  optimization, decompressing a block with 1 long match goes
  from 5.7 GB/s to 800 MB/s.
* Refactor `ZSTD_execSequenceEnd()`.
* Increase the literal copy shortcut to 16.
* Add a shortcut for offset >= 16.
* Simplify `ZSTD_execSequence()` by pushing more cases into
  `ZSTD_execSequenceEnd()`.
* Delete `ZSTD_execSequenceLong()` since it is exactly the
  same as `ZSTD_execSequence()`.
* Refactor and fix `ZSTD_storeSeq()` to remove the implicit requirement that the literals can be wildcopied.

| Scenario | gcc-9 | clang-8 |
|---|---|---|
| Decompression | +12.4% | +22.8% |
| Level 1 | +3.8% | +5.0% |
| Level 2 | -0.9% | +4.2% |
| Level 3 | +2.4% | +2.7% |
| Level 4 | +1.8% | +3.1% |
| Level 5 | +3.3% | +2.0% |

The numbers are gathered on an Intel i9-9900k with turbo disabled taskset to CPU 0. It is the average percent change over silesia.tar, enwik8, and calgary.tar. The decompression speed is the average of the percent change over levels 1 to 5.

I got very similar results with the fbcode gcc and clang on my devserver for compression and decompression. On OS X gcc sees a 9% decompression speed boost, and clang a 13% speed boost, and compression is neutral.

Edge case single long match:
The file `head -c 128K /dev/zero` triggers a corner case in the new code. Level 1 compression is now 12.7 GB/s (previously 8.8 GB/s), and decompression is neutral at 5.8 GB/s.

Crdit to OSS-Fuzz for finding the wildcopy overread.