@Cyan4973 2019-09-20
That's a hell of a patch @terrelln !

Quick results on my bench system (Core i7-9700k, gcc 8.3.0) :

silesia : 1312 -> 1400
calgary : 1189 -> 1288
enwik8 : 1059 -> 1152

Slightly less than 10%, still excellent !
I also tested the specific case of `--ultra -22`, since it triggers long distance matches, with resulting caches misses (compensated with prefetching) : 

silesia -22 : 1118 -> 1243
enwik9 -22 : 865 -> 975

This time, this is more than 10% !

With clang 8.0, as announced, the gains are much larger : 

silesia : 1082 -> 1268
calgary : 963 -> 1144
enwik8 : 854 -> 1016

Congratulations !!