@davidbolvansky 2019-09-20
And if you remove "dont vectorize" macro and enable -ftree-vectorize again in Makefile, you should see another small again. Seen with gcc9.