@davidbolvansky 2019-09-20
Consider 
COPY32(op, ip);

instead of 
COPY16(op, ip);
COPY16(op, ip);

Brings another perf gain.