@davidbolvansky 2019-09-20
zstd's master + GCC 9
./zstd -b -d bench/nci.zst -f
 3#nci.zst           :  33553445 ->   2857582 (11.74),  0.00 MB/s ,1295.1 MB/s

This patch + things I mentioned + unrolled building loop in ZSTD_buildFSETable + GCC 9

/zstd -b -d bench/nci.zst -f
 3#nci.zst           :  33553445 ->   2857582 (11.74),  0.00 MB/s ,1645.9 MB/s 

22% percent improvement