@terrelln 2019-09-20
Looks like a new assert in `ZSTD_storeSeq()` uncovered an existing bug where there wasn't at least 8 bytes after the literals length for wildcopy.