@Cyan4973 2019-09-20
Indeed, the end of buffer condition was previously implicit.

nit-picking here, but the exact condition is that end of buffer must be at least 8 bytes beyond _start_ of literals, and at least **7** bytes beyond end of literals. I don't think it makes a difference but just in case it does, I mention it here.

The implicit end of buffer condition relied on the fact that the last match starts effectively at least 8 bytes before end of block, due to hashing constraints. `repcode` can evade the hashing issue, but I think the same condition is applied to them anyway.

The new code in `decodeCorpus` doesn't respect this condition, since end of buffer is declared right at `literals + litlen`, so without sufficient margin : 
https://github.com/facebook/zstd/pull/1804/files#diff-771ac86c47832895df8cc02374f8da6cR761

However, the errors detected in `travisCI` do not depend on `decodeCorpus`, so there is something else going on.