@terrelln 2019-09-20
LDM also doesn't respect this condition, which is where the errors are showing up.

I'm working on rewriting `ZSTD_storeSeq()` to not require any over length, since that cleans up implicit assumptions.