@Cyan4973 2019-09-20
Yes, this version is even better.
Testing at level 1, which is likely the most impacted by `ZSTD_storeSeq()`, I now measure a _gain_ in compression speed : 

silesia : 498 -> 523
calgary : 417 -> 434
enwik8 : 365 -> 378

I would even say, the benefits are larger than I expected.

Decompression speed also improves a little bit more compared to latest measurement.