@terrelln 2019-09-20
@davidbolvansky I haven't been able to reproduce speed gains of removing `DONT_VECTORIZE` and using `COPY32()` with gcc-9 on an Intel i9-9900K. I will leave them out of this patch, but if you put up a PR with those changes I'll benchmark them.

What do you mean by `ZSTD_buildFSETable()` unrolling, are you planning on putting a PR up for that?