@terrelln 2019-09-20
I've fixed the compression perf regression on OS X clang by using `HINT_INLINE` for `ZSTD_storeSeq()`. Clang on Mac wasn't inlining it for level 1...

I've verified that it doesn't change the performance of anything else. I've been running fuzzers for a few hours, and no failures so far. I see good performance on my home server, devserver, and Macbook (detailed in summary).

I'm happy this this code now.