@Cyan4973 2019-09-21
Some minor detail left : 
https://travis-ci.org/facebook/zstd/jobs/587703350#L3968

`HINT_INLINE` needs an additional attribute to avoid triggering the `unused-function` warning.