@davidbolvansky 2019-09-21
> I will measure (my) things again after this patch is merged.

With this patch:
 ./zstd -b -d bench/nci.zst -f 
 3#nci.zst           :  33553445 ->   2857582 (11.74),  0.00 MB/s ,1611.7 MB/s 

Some my changes:
 ./zstd -b -d bench/nci.zst -f 
 3#nci.zst           :  33553445 ->   2857582 (11.74),  0.00 MB/s ,1632.9 MB/s 

https://pastebin.com/3uumCPrf